- Combined APIs for simplicity.
