+++ src/app/api/[type]/[id]/[version]/alloy/[alloyName]/route.ts
@@ -1,19 +1,66 @@
-import { Alloy } from "@/types";
-import { NextResponse } from "next/server";
-import alloys from "@/data/alloys.json";
+import {Alloy, Mineral, MineralUse} from "@/types";
+import {NextResponse} from "next/server";
+import alloysJson from "@/data/alloys.json";
+import mineralsJson from "@/data/minerals.json";
+
 
 export async function GET(
-		_ : Request,
-		{ params }: { params: { alloyName: string } }) {
-	const alloyName = params.alloyName;
+		request : Request,
+		{params} : { params : { alloyName : string } }
+) {
+	const {alloyName} = params;
+	const {searchParams} = new URL(request.url);
+	const uses = searchParams.getAll("uses");
 
-	const alloyData = (alloys as { alloys: Alloy[] }).alloys.find(
-			(alloy) => alloy.name.toLowerCase() === decodeURIComponent(alloyName).toLowerCase()
+	const alloyList = alloysJson.alloys as Alloy[];
+	const alloy = alloyList.find(
+			(a) => a.name.toLowerCase() === decodeURIComponent(alloyName).toLowerCase()
 	);
 
-	if (!alloyData) {
-		return NextResponse.json({ error: "Alloy not found" }, { status: 404 });
+	if (!alloy) {
+		return NextResponse.json({message : "Alloy not found"}, {status : 404});
 	}
 
-	return NextResponse.json(alloyData);
+	const alloyMinerals = alloy.components.flatMap((component) => {
+		const rawMinerals = mineralsJson[component.mineral as keyof typeof mineralsJson];
+
+		if (!rawMinerals) {
+			return null;
+		}
+
+		const mineralsWithProduces : Mineral[] = rawMinerals.map((mineral) => ({
+			...mineral,
+			produces : component.mineral,
+			uses : mineral.uses ? toMineralUses(mineral.uses) : undefined
+		}));
+
+		return mineralsWithProduces.filter((mineral) => {
+			if (uses.length === 0) {
+				return true;
+			}
+
+			return uses.some((use) => mineral.uses?.includes(use as MineralUse));
+		});
+	}).filter((m) : m is NonNullable<typeof m> => m !== null);
+
+	const response = {
+		alloy : {
+			name : alloy.name,
+			components : alloy.components
+		},
+		minerals : alloyMinerals
+	};
+
+	return NextResponse.json(response);
+}
+
+/**
+ * Utility function to convert string mineral uses to enums.
+ * @param uses array of mineral uses as strings.
+ */
+function toMineralUses(uses : string[]) : MineralUse[] {
+	const validUses = Object.values(MineralUse) as string[];
+	return uses
+			.filter((use) => validUses.includes(use))
+			.map((use) => use as MineralUse);
 }
